Updates based on testing
Removed extra checkSelectAll as it is also called via another method.
For some reason the duplicate was resulting in the first letter entered
being removed.  Also changed min width back to inherit.  This will
cause the box-shadow to reappear, but allows the element to size
properly when being used in the date selector.